main,refactor: move addTotals() from main.c to newly introduce source file
