+++ main/general.h
@@ -18,6 +18,19 @@
 # include "e_msoft.h"
 #endif
 
+/*  To provide timings features if available.
+ */
+#ifdef HAVE_CLOCK
+# ifdef HAVE_TIME_H
+#  include <time.h>
+# endif
+#else
+# ifdef HAVE_TIMES
+#  ifdef HAVE_SYS_TIMES_H
+#   include <sys/times.h>
+#  endif
+# endif
+#endif
 
 /*
 *   MACROS
@@ -75,6 +88,28 @@ extern char *getenv (const char *);
 #define iconv libiconv
 #define iconv_open libiconv_open
 #define iconv_close libiconv_close
+#endif
+
+/*
+*  Prepare clock() and its related macros
+*/
+#if defined (HAVE_CLOCK)
+# define CLOCK_AVAILABLE
+# ifndef CLOCKS_PER_SEC
+#  define CLOCKS_PER_SEC		1000000
+# endif
+#elif defined (HAVE_TIMES)
+# define CLOCK_AVAILABLE
+# define CLOCKS_PER_SEC	60
+static clock_t clock (void)
+{
+	struct tms buf;
+
+	times (&buf);
+	return (buf.tms_utime + buf.tms_stime);
+}
+#else
+# define clock()  (clock_t)0
 #endif
 
 #endif  /* CTAGS_MAIN_GENERAL_H */
