+++ main/main.c
@@ -30,20 +30,6 @@
 
 #include <string.h>
 
-/*  To provide timings features if available.
- */
-#ifdef HAVE_CLOCK
-# ifdef HAVE_TIME_H
-#  include <time.h>
-# endif
-#else
-# ifdef HAVE_TIMES
-#  ifdef HAVE_SYS_TIMES_H
-#   include <sys/times.h>
-#  endif
-# endif
-#endif
-
 /*  To provide directory searching for recursion feature.
  */
 
@@ -72,6 +58,7 @@
 #include "parse_p.h"
 #include "read_p.h"
 #include "routines_p.h"
+#include "stats_p.h"
 #include "trace.h"
 #include "trashbox_p.h"
 #include "writer_p.h"
@@ -84,14 +71,8 @@
 #endif
 
 /*
-*   MACROS
-*/
-#define plural(value)  (((unsigned long)(value) == 1L) ? "" : "s")
-
-/*
 *   DATA DEFINITIONS
 */
-static struct { long files, lines, bytes; } Totals = { 0, 0, 0 };
 static mainLoopFunc mainLoop;
 static void *mainData;
 
@@ -103,15 +84,6 @@ static bool createTagsForEntry (const char *const entryName);
 /*
 *   FUNCTION DEFINITIONS
 */
-
-extern void addTotals (
-		const unsigned int files, const long unsigned int lines,
-		const long unsigned int bytes)
-{
-	Totals.files += files;
-	Totals.lines += lines;
-	Totals.bytes += bytes;
-}
 
 extern bool isDestinationStdout (void)
 {
@@ -355,69 +327,6 @@ static bool createTagsFromListFile (const char *const fileName)
 	return resize;
 }
 
-#if defined (HAVE_CLOCK)
-# define CLOCK_AVAILABLE
-# ifndef CLOCKS_PER_SEC
-#  define CLOCKS_PER_SEC		1000000
-# endif
-#elif defined (HAVE_TIMES)
-# define CLOCK_AVAILABLE
-# define CLOCKS_PER_SEC	60
-static clock_t clock (void)
-{
-	struct tms buf;
-
-	times (&buf);
-	return (buf.tms_utime + buf.tms_stime);
-}
-#else
-# define clock()  (clock_t)0
-#endif
-
-static void printTotals (const clock_t *const timeStamps)
-{
-	const unsigned long totalTags = numTagsTotal();
-	const unsigned long addedTags = numTagsAdded();
-
-	fprintf (stderr, "%ld file%s, %ld line%s (%ld kB) scanned",
-			Totals.files, plural (Totals.files),
-			Totals.lines, plural (Totals.lines),
-			Totals.bytes/1024L);
-#ifdef CLOCK_AVAILABLE
-	{
-		const double interval = ((double) (timeStamps [1] - timeStamps [0])) /
-								CLOCKS_PER_SEC;
-
-		fprintf (stderr, " in %.01f seconds", interval);
-		if (interval != (double) 0.0)
-			fprintf (stderr, " (%lu kB/s)",
-					(unsigned long) (Totals.bytes / interval) / 1024L);
-	}
-#endif
-	fputc ('\n', stderr);
-
-	fprintf (stderr, "%lu tag%s added to tag file",
-			addedTags, plural(addedTags));
-	if (Option.append)
-		fprintf (stderr, " (now %lu tags)", totalTags);
-	fputc ('\n', stderr);
-
-	if (totalTags > 0  &&  Option.sorted != SO_UNSORTED)
-	{
-		fprintf (stderr, "%lu tag%s sorted", totalTags, plural (totalTags));
-#ifdef CLOCK_AVAILABLE
-		fprintf (stderr, " in %.02f seconds",
-				((double) (timeStamps [2] - timeStamps [1])) / CLOCKS_PER_SEC);
-#endif
-		fputc ('\n', stderr);
-	}
-
-#ifdef DEBUG
-	fprintf (stderr, "longest tag line = %lu\n",
-		 (unsigned long) maxTagsLine ());
-#endif
-}
-
 static bool etagsInclude (void)
 {
 	return (bool)(Option.etags && Option.etagsInclude != NULL);
@@ -482,7 +391,7 @@ static void batchMakeTags (cookedArgs *args, void *user CTAGS_ATTR_UNUSED)
 	timeStamp (2);
 
 	if (Option.printTotals)
-		printTotals (timeStamps);
+		printTotals (timeStamps, Option.append, Option.sorted);
 #undef timeStamp
 }
