+++ main/parse.c
@@ -27,7 +27,6 @@
 #include "htable.h"
 #include "keyword.h"
 #include "lxpath_p.h"
-#include "main_p.h"
 #include "param.h"
 #include "param_p.h"
 #include "parse_p.h"
@@ -40,6 +39,7 @@
 #include "read_p.h"
 #include "routines.h"
 #include "routines_p.h"
+#include "stats_p.h"
 #include "subparser.h"
 #include "subparser_p.h"
 #include "trace.h"
