+++ main/read.c
@@ -22,12 +22,12 @@
 #include "read_p.h"
 #include "debug.h"
 #include "entry_p.h"
-#include "main_p.h"
 #include "routines.h"
 #include "routines_p.h"
 #include "options_p.h"
 #include "parse_p.h"
 #include "promise_p.h"
+#include "stats_p.h"
 #include "trace.h"
 #include "trashbox.h"
 #ifdef HAVE_ICONV
